renamed assoc to set
