+++ src/Dict/index.ts
@@ -40,13 +40,13 @@ export declare function merge<A, B>(snd: B): (fst: A) => A & B
 
 /** Adds a property. Equivalent to merging with `{key: value}` */
 
-export declare function assoc<T, K extends string, V>(
+export declare function set<T, K extends string, V>(
   dict: T,
   key: K,
   value: V,
 ): T & { [key in K]: V }
 
-export declare function assoc<T, K extends string, V>(
+export declare function set<T, K extends string, V>(
   key: K,
   value: V,
 ): (dict: T) => T & { [key in K]: V }
